requisition-service/requisition_service/serializers.py
from rest_framework import serializers
from .models import Requisition

class RequisitionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Requisition
        fields = '__all__'
